Refactoring: GridCreatingConditionTriangle::RemovePolyLineVertexCommand implementation is separated from gridcreatingconditiontriangle.cpp.
