+++ libs/gridcreatingcondition/triangle/gridcreatingconditiontriangle.cpp
@@ -20,6 +20,7 @@
 #include "private/gridcreatingconditiontriangle_movepolylinecommand.h"
 #include "private/gridcreatingconditiontriangle_movepolylinevertexcommand.h"
 #include "private/gridcreatingconditiontriangle_removepolygonvertexcommand.h"
+#include "private/gridcreatingconditiontriangle_removepolylinevertexcommand.h"
 
 #include <guibase/widget/waitdialog.h>
 #include <guicore/base/iricmainwindowinterface.h>
@@ -334,76 +335,6 @@ void GridCreatingConditionTriangle::mouseMoveEvent(QMouseEvent* event, PreProces
 		break;
 	}
 }
-
-class GridCreatingConditionTriangle::RemovePolyLineVertexCommand : public QUndoCommand
-{
-public:
-	RemovePolyLineVertexCommand(vtkIdType vertexId, GridCreatingConditionTriangle* pol) :
-		QUndoCommand {GridCreatingConditionTriangle::tr("Remove Break Line Vertex")}
-	{
-		m_vertexId = vertexId;
-		double p[3];
-		pol->m_selectedLine->getVtkLine()->GetPoints()->GetPoint(m_vertexId, p);
-		m_vertexPosition = QVector2D(p[0], p[1]);
-		m_polygon = pol;
-		m_targetLine = m_polygon->m_selectedLine;
-	}
-	void redo() {
-		vtkPoints* points = m_targetLine->getVtkLine()->GetPoints();
-		QVector<QVector2D> positions;
-		positions.reserve(points->GetNumberOfPoints());
-		double p[3];
-		for (vtkIdType i = 0; i < m_vertexId; ++i) {
-			points->GetPoint(i, p);
-			positions.append(QVector2D(p[0], p[1]));
-		}
-		// skip vertex in m_vertexId[
-		for (vtkIdType i = m_vertexId + 1; i < points->GetNumberOfPoints(); ++i) {
-			points->GetPoint(i, p);
-			positions.append(QVector2D(p[0], p[1]));
-		}
-		points->SetNumberOfPoints(positions.count());
-		for (vtkIdType i = 0; i < positions.count(); ++i) {
-			QVector2D v = positions.at(i);
-			points->SetPoint(i, v.x(), v.y(), 0);
-		}
-		points->Modified();
-		m_polygon->m_mouseEventMode = GridCreatingConditionTriangle::meNormal;
-		m_targetLine->getVtkLine()->Modified();
-		m_targetLine->updateShapeData();
-		m_polygon->renderGraphicsView();
-	}
-	void undo() {
-		vtkPoints* points = m_targetLine->getVtkLine()->GetPoints();
-		QVector<QVector2D> positions;
-		positions.reserve(points->GetNumberOfPoints());
-		double p[3];
-		for (vtkIdType i = 0; i < m_vertexId; ++i) {
-			points->GetPoint(i, p);
-			positions.append(QVector2D(p[0], p[1]));
-		}
-		positions.append(m_vertexPosition);
-		for (vtkIdType i = m_vertexId; i < points->GetNumberOfPoints(); ++i) {
-			points->GetPoint(i, p);
-			positions.append(QVector2D(p[0], p[1]));
-		}
-		points->SetNumberOfPoints(positions.count());
-		for (vtkIdType i = 0; i < positions.count(); ++i) {
-			QVector2D v = positions.at(i);
-			points->SetPoint(i, v.x(), v.y(), 0);
-		}
-		points->Modified();
-		m_targetLine->getVtkLine()->Modified();
-		m_targetLine->updateShapeData();
-		m_polygon->renderGraphicsView();
-	}
-
-private:
-	vtkIdType m_vertexId;
-	QVector2D m_vertexPosition;
-	GridCreatingConditionTriangle* m_polygon;
-	GridCreatingConditionTriangleAbstractLine* m_targetLine;
-};
 
 void GridCreatingConditionTriangle::mousePressEvent(QMouseEvent* event, PreProcessorGraphicsViewInterface* v)
 {
